+++ routes/stores.js
@@ -75,8 +75,8 @@ router.post('/transactions', (req, res) => {
     let store_id = parseInt(((req.headers.referer).split('/'))[4]),
         min_price = parseFloat(req.body.min_price),
         max_price = parseFloat(req.body.max_price),
-        min_pieces = parseInt(req.body.min_pieces),
-        max_pieces = parseInt(req.body.max_pieces),
+        min_pieces = parseFloat(req.body.min_pieces),
+        max_pieces = parseFloat(req.body.max_pieces),
         payment_method = req.body.payment_method
 
     if (payment_method === '') payment_method = 'All'
